feat: add disambiguator class
